Refactor waist measurement
