+++ src/Measurements.Core/Bones.cs
@@ -2,21 +2,5 @@ namespace Measurements
 {
     internal enum Bones
     {
-        /// <summary>
-        /// Waist (front)
-        /// </summary>
-        N_Waist_f,
-        /// <summary>
-        /// Waist (back)
-        /// </summary>
-        N_Waist_b,
-        /// <summary>
-        /// Waist (left)
-        /// </summary>
-        N_Waist_L,
-        /// <summary>
-        /// Waist (right)
-        /// </summary>
-        N_Waist_R,
     }
 }
