+++ src/Measurements.Core/MeasurementsGui.cs
@@ -14,7 +14,7 @@ namespace Measurements
             new Height.Gui(),
             new BraSize.Gui(),
             new Bust.Gui(),
-            new Gui.WaistGui(),
+            new Waist.Gui(),
             new Hips.Gui(),
             new Gui.WaistToHipsGui(),
             new Dick.Gui(),
