+++ src/Measurements.Core/Waist/Bones.cs
@@ -1,0 +1,22 @@
+namespace Measurements.Waist
+{
+    internal enum Bones
+    {
+        /// <summary>
+        /// Waist (front)
+        /// </summary>
+        N_Waist_f,
+        /// <summary>
+        /// Waist (back)
+        /// </summary>
+        N_Waist_b,
+        /// <summary>
+        /// Waist (left)
+        /// </summary>
+        N_Waist_L,
+        /// <summary>
+        /// Waist (right)
+        /// </summary>
+        N_Waist_R,
+    }
+}
